# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/b1/subaccounts": {
            "get": {
                "tags": [
                    "BeOne Subaccounts"
                ],
                "summary": "Get all BeOne subaccounts",
                "description": "Get all BeOne subaccounts",
                "operationId": "1cda098f013ff4a1b4c909331cc7c382",
                "responses": {
                    "200": {
                        "description": "Subaccounts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subaccounts": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "subaccount_id": {
                                                        "description": "Subaccount UUID",
                                                        "type": "string",
                                                        "example": "123456"
                                                    },
                                                    "external_id": {
                                                        "description": "Subaccount ExternalId",
                                                        "type": "string",
                                                        "example": "123456"
                                                    },
                                                    "description": {
                                                        "description": "Subaccount Description",
                                                        "type": "string",
                                                        "example": "My subaccount"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message",
                                            "type": "string",
                                            "example": "Unauthorized | Error while decoding the token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/b1/subaccount": {
            "post": {
                "tags": [
                    "BeOne Subaccounts"
                ],
                "summary": "Create a new BeOne subaccount",
                "description": "Create a new BeOne subaccount",
                "operationId": "f22d0bd86a8a574e8b639408ec954f16",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Description"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "My subaccount"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subaccount created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subaccount_id": {
                                            "description": "Subaccount UUID",
                                            "type": "string",
                                            "example": "123456"
                                        },
                                        "external_id": {
                                            "description": "Subaccount ExternalId",
                                            "type": "string",
                                            "example": "123456"
                                        },
                                        "description": {
                                            "description": "Subaccount Description",
                                            "type": "string",
                                            "example": "My subaccount"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message",
                                            "type": "string",
                                            "example": "Unauthorized | Error while decoding the token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/b1/subaccount/{uuid}": {
            "get": {
                "tags": [
                    "BeOne Subaccounts"
                ],
                "summary": "Get a specific BeOne subaccount",
                "description": "Get a specific BeOne subaccount by UUID",
                "operationId": "0531cdf21f1a95e0b5fce0e29611af05",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Subaccount UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subaccount retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subaccount_id": {
                                            "description": "Subaccount UUID",
                                            "type": "string",
                                            "example": "123456"
                                        },
                                        "external_id": {
                                            "description": "Subaccount ExternalId",
                                            "type": "string",
                                            "example": "123456"
                                        },
                                        "description": {
                                            "description": "Subaccount Description",
                                            "type": "string",
                                            "example": "My subaccount"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message",
                                            "type": "string",
                                            "example": "Unauthorized | Error while decoding the token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subaccount not found or you do not have permission to access it",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message",
                                            "type": "string",
                                            "example": "Subaccount not found or you do not have permission to access it"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/b1/subaccount/{uuid}/balance": {
            "get": {
                "tags": [
                    "BeOne Subaccounts"
                ],
                "summary": "Get the balance of a specific BeOne subaccount",
                "description": "Get the balance of a specific BeOne subaccount by UUID",
                "operationId": "d98539cbd0ec94d8fd1609445214a411",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Subaccount UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Balance retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "balance": {
                                            "description": "Subaccount balance",
                                            "type": "number",
                                            "example": "100.00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/b1/subaccount/{uuid}/movements": {
            "get": {
                "tags": [
                    "BeOne Subaccounts"
                ],
                "summary": "Get all movements for a specific subaccount",
                "description": "Get all movements for a specific subaccount",
                "operationId": "f5ca334e0a18cd51a899d60f27e92a4b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Subaccount UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123456"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "from": {
                                        "description": "From date (Unix Timestamp)",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "to": {
                                        "description": "To date (Unix Timestamp)",
                                        "type": "string",
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Movements retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "movements": {
                                            "description": "Movements",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "movement_id": {
                                                        "description": "Movement UUID",
                                                        "type": "string",
                                                        "example": "123456"
                                                    },
                                                    "date": {
                                                        "description": "Movement Date / Unix Timestamp",
                                                        "type": "string",
                                                        "example": "1716611739"
                                                    },
                                                    "type": {
                                                        "description": "Movement Type",
                                                        "type": "string",
                                                        "example": "deposit"
                                                    },
                                                    "amount": {
                                                        "description": "Movement Amount",
                                                        "type": "string",
                                                        "example": "100.00"
                                                    },
                                                    "authorization_code": {
                                                        "description": "Authorization Code",
                                                        "type": "string",
                                                        "example": "123456"
                                                    },
                                                    "description": {
                                                        "description": "Movement Description",
                                                        "type": "string",
                                                        "example": "Deposit"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total_records": {
                                            "description": "Total records",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "from": {
                                            "description": "From date (Unix Timestamp)",
                                            "type": "string",
                                            "example": "1234567890"
                                        },
                                        "to": {
                                            "description": "To date (Unix Timestamp)",
                                            "type": "string",
                                            "example": "1234567890"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}